Refactor REST mapping to use GVK objects for improved type safety and clarity (#14)
This pull request introduces several key changes to the `cloudcoil`
client configuration and resource management. The primary focus is on
refactoring the handling of GroupVersionKind (GVK) to improve code
clarity and maintainability. The most important changes include the
introduction of a new `GVK` class, modifications to the `Resource` class
to use this new class, and updates to the client configuration to
leverage the new `GVK` structure.

### Refactoring to use GVK:

*
[`cloudcoil/client/_config.py`](diffhunk://#diff-0794500b70c79f233e6f3cb70a9dec87baf56a0626947b46653c915d034dd472L15-R15):
Updated imports to include `GVK` from `cloudcoil.client._resource` and
modified the `_rest_mapping` dictionary to use `GVK` as the key instead
of a tuple of strings.
[[1]](diffhunk://#diff-0794500b70c79f233e6f3cb70a9dec87baf56a0626947b46653c915d034dd472L15-R15)
[[2]](diffhunk://#diff-0794500b70c79f233e6f3cb70a9dec87baf56a0626947b46653c915d034dd472L140-R140)
[[3]](diffhunk://#diff-0794500b70c79f233e6f3cb70a9dec87baf56a0626947b46653c915d034dd472L201-R201)
[[4]](diffhunk://#diff-0794500b70c79f233e6f3cb70a9dec87baf56a0626947b46653c915d034dd472L218-R234)

### Introduction of GVK class:

*
[`cloudcoil/client/_resource.py`](diffhunk://#diff-87a7c23e2c8e571f89d57831b737bc1cd7f68b0bd85f7f520f925b57554e26d1R11-R44):
Added a new `GVK` class that encapsulates `api_version` and `kind`, and
includes properties for `group` and `version`. This class is used to
replace the tuple of strings previously used for GVK representation.

### Enhancements to Resource class:

*
[`cloudcoil/client/_resource.py`](diffhunk://#diff-87a7c23e2c8e571f89d57831b737bc1cd7f68b0bd85f7f520f925b57554e26d1R54-R65):
Modified the `Resource` class to use the new `GVK` class in the `gvk`
method. Added `name` and `namespace` properties to the `Resource` class
for easier access to metadata.