Add update and fetch methods for resource management (#15)
This pull request includes several changes to enhance the functionality
of namespace management and improve the handling of API responses. The
most important changes include adding update and fetch methods for
namespaces, modifying the API client to handle updates, and adding tests
for the new functionality.

Enhancements to namespace management:

*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5R55-R65):
Added examples to demonstrate how to modify and update a namespace, as
well as how to fetch the latest version of a namespace from the server.
*
[`cloudcoil/client/_resource.py`](diffhunk://#diff-87a7c23e2c8e571f89d57831b737bc1cd7f68b0bd85f7f520f925b57554e26d1L76-R116):
Added `fetch`, `update`, `async_fetch`, and `async_update` methods to
the `Resource` class to support fetching and updating resources.
*
[`cloudcoil/client/_resource.py`](diffhunk://#diff-87a7c23e2c8e571f89d57831b737bc1cd7f68b0bd85f7f520f925b57554e26d1R60-R79):
Added setters for `name` and `namespace` properties to allow setting
these attributes directly.

Improvements to API client:

*
[`cloudcoil/client/_api_client.py`](diffhunk://#diff-003c287eabfb30f1d5ac7eab59dcc0603f8d6e6c45e7554828113dfffcdb8066L72-R75):
Modified the `create` method to remove the `namespace` parameter and use
the namespace from the body. Added `update` and `async_update` methods
to handle updates.
[[1]](diffhunk://#diff-003c287eabfb30f1d5ac7eab59dcc0603f8d6e6c45e7554828113dfffcdb8066L72-R75)
[[2]](diffhunk://#diff-003c287eabfb30f1d5ac7eab59dcc0603f8d6e6c45e7554828113dfffcdb8066R85-R98)
[[3]](diffhunk://#diff-003c287eabfb30f1d5ac7eab59dcc0603f8d6e6c45e7554828113dfffcdb8066R171-R184)
*
[`cloudcoil/client/_api_client.py`](diffhunk://#diff-003c287eabfb30f1d5ac7eab59dcc0603f8d6e6c45e7554828113dfffcdb8066L48-R48):
Updated the `_handle_create_response` method to use
`response.is_success` instead of checking for specific status codes.

Tests for new functionality:

*
[`tests/test_e2e.py`](diffhunk://#diff-ed3b8af48cd21dc1ac735b839d89b172a6f45569c324af70a2bdb4f60e5fbcddR16-R18):
Added tests to verify that namespaces can be updated and fetched
correctly, including both synchronous and asynchronous operations.
[[1]](diffhunk://#diff-ed3b8af48cd21dc1ac735b839d89b172a6f45569c324af70a2bdb4f60e5fbcddR16-R18)
[[2]](diffhunk://#diff-ed3b8af48cd21dc1ac735b839d89b172a6f45569c324af70a2bdb4f60e5fbcddR34-R36)