Add scheme to parse unknown objects (#18)
This pull request includes several changes to the `cloudcoil` project,
focusing on enhancing the resource parsing functionality and updating
the documentation. The most important changes include the addition of a
new `Scheme` class for resource management, updates to the `GVK` and
`BaseResource` classes, and improvements to the `README.md` file for
better user guidance.

Enhancements to resource parsing:

*
[`cloudcoil/client/_resource.py`](diffhunk://#diff-87a7c23e2c8e571f89d57831b737bc1cd7f68b0bd85f7f520f925b57554e26d1L2-R2):
Updated the `GVK` and `BaseResource` classes to use `Annotated` types
and `Field` from `pydantic` for better alias handling.
[[1]](diffhunk://#diff-87a7c23e2c8e571f89d57831b737bc1cd7f68b0bd85f7f520f925b57554e26d1L2-R2)
[[2]](diffhunk://#diff-87a7c23e2c8e571f89d57831b737bc1cd7f68b0bd85f7f520f925b57554e26d1L11-R18)
[[3]](diffhunk://#diff-87a7c23e2c8e571f89d57831b737bc1cd7f68b0bd85f7f520f925b57554e26d1L32-R33)
*
[`cloudcoil/scheme.py`](diffhunk://#diff-e5e93640934f2c889c3397733465827f1103a90ad8db64e725472a2017c39790R1-R82):
Introduced a new `Scheme` class to handle resource registration,
discovery, and parsing. This class includes methods for registering
resources, discovering resources in modules, and parsing YAML files into
resource objects.

Documentation improvements:

*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5R82-R105):
Added a new section demonstrating how to parse Kubernetes resources
using the `Scheme` class and create a job from a YAML file.
*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5R31-R32):
Included an import statement for `Path` from `pathlib` in the Quick
Start section.

Codebase updates:

*
[`cloudcoil/client/__init__.py`](diffhunk://#diff-4057ef0fe8860290f3edae2ccc8c19ded5c3e67bdebbe978ef5a210be18b0a53L3-R5):
Added `GVK` to the list of exports to make it available for import from
the `cloudcoil.client` module.