Improve parser to handle model dedupe (#32)
This pull request includes significant changes to the code generation
process, improvements to testing, and updates to the `Makefile`. The
most important changes include replacing the custom script for
generating models with a new tool, enhancing the testing setup, and
simplifying the import rewriting process.

### Code Generation Improvements:
*
[`Makefile`](diffhunk://#diff-76ed074a9305c04054cdebb9e9aad2d818052b07091de1f20cad0bbac34ffb52L38-R38):
Replaced the custom script `apimachinery_gen.py` with the new
`cloudcoil-model-codegen` tool for generating models.
*
[`cloudcoil/apimachinery.py`](diffhunk://#diff-b77e95e9e9ee74eb756ff07bdcdf7a96b1e127f86589c0230f621e711c4379e3L1-R2):
Updated the header to indicate the file is generated by
`cloudcoil-model-codegen` and added a "DO NOT EDIT" notice.
*
[`scripts/apimachinery_gen.py`](diffhunk://#diff-7ec409e62096f68d9c7de2cc4b04334337bbb29e055c06fa62813e03f895cfeaL1-L97):
Removed the custom script used for generating models from the OpenAPI
spec.

### Testing Enhancements:
*
[`pyproject.toml`](diffhunk://#diff-50c86b7ed8ac2cf95bd48334961bf0530cdc77b5a56f852c5c61b89d735fd711L62-R66):
Added `pytest-xdist` to the development dependencies for parallel test
execution and updated pytest options to run tests in parallel.
[[1]](diffhunk://#diff-50c86b7ed8ac2cf95bd48334961bf0530cdc77b5a56f852c5c61b89d735fd711L62-R66)
[[2]](diffhunk://#diff-50c86b7ed8ac2cf95bd48334961bf0530cdc77b5a56f852c5c61b89d735fd711R85)
*
[`cloudcoil/_testing/plugin.py`](diffhunk://#diff-f6fd8eb2539c3388dfa73979bd7f2fda0a8a859179cc45b9f18126eff46f40dcL48-R63):
Improved the `test_cluster` function to handle potential race conditions
when writing the k3d binary.

### Import Rewriting:
*
[`cloudcoil/codegen/import_rewriter.py`](diffhunk://#diff-e3dc1ac49f81268e32a3a3119c546bf4071f0f1f695324642bb3f57d5e48ff8fR1-R92):
Added a new script to handle the rewriting of imports in generated
files, ensuring they are correctly referenced within the `cloudcoil`
package.

### Miscellaneous:
*
[`tests/test_codegen.py`](diffhunk://#diff-1582e084856121640b055f625172ffcfef41bb645dc43ba7abf8b9eb161e47e1L5-R5):
Updated import paths to reflect the new location of the code generation
functions.