Add support for exec auth (#44)
This pull request introduces an `ExecAuthenticator` class to handle
Kubernetes exec authentication and includes corresponding updates to the
configuration and testing files. The most important changes are grouped
by theme as follows:

### New Authentication Class:
* Added `ExecAuthenticator` class to handle exec-based authentication in
the `cloudcoil/client/_config.py` file. This class executes a command to
retrieve an authentication token and caches it until it expires.

### Configuration Updates:
* Updated the `Auth` type alias and modified the `__init__` method in
`cloudcoil/client/_config.py` to use the new `ExecAuthenticator` class
for exec-based authentication.
[[1]](diffhunk://#diff-0794500b70c79f233e6f3cb70a9dec87baf56a0626947b46653c915d034dd472L32-R101)
[[2]](diffhunk://#diff-0794500b70c79f233e6f3cb70a9dec87baf56a0626947b46653c915d034dd472L97-R160)

### Testing Enhancements:
* Added a new test case `test_exec_auth` in `tests/test_config.py` to
validate the behavior of the `ExecAuthenticator` class, including the
execution of the command and the handling of the returned token.

Signed-off-by: Sambhav Kothari <sambhavs.email@gmail.com>