Add support for dynamic resources (#46)
This pull request includes updates to the `README.md` file to enhance
the documentation for dynamic resource handling and resource parsing in
the `cloudcoil` library. The most important changes include adding a new
section on dynamic resources, updating examples for resource parsing,
and correcting the import statement for type checking.

Enhancements to documentation:

*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5R134-R168):
Added a new section "Dynamic Resources" with an example of using the
`get_dynamic_resource` function to create, update, and access Kubernetes
resources dynamically.

Updates to resource parsing examples:

*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L145-R180):
Updated the example for getting a resource class by GVK to clarify that
it applies to existing resource model classes.

Corrections to type checking examples:

*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L275-R313):
Corrected the import statement in the type checking example to use
`resources` instead of `scheme` for getting the model class.