Add velero models and sealed-secrets to extras (#82)
This pull request introduces a new integration for Velero, along with
relevant documentation and configuration updates. The most important
changes include adding Velero to the list of first-class integrations,
creating a new README for Velero with installation and usage examples,
and updating the configuration files to support the new integration.

### New Velero Integration:

*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5R67):
Added Velero to the list of first-class integrations.

### Documentation:

*
[`models/velero/README.md`](diffhunk://#diff-91a184c8cb1e489edc6c08fb5ef73d496172c74b5e77c782be9444a60f1cd08fR1-R106):
Created a new README for Velero with installation instructions, usage
examples, and documentation links.

### Configuration:

*
[`models/velero/cookiecutter.yaml`](diffhunk://#diff-9d319c20c6957e36569922e037cce7df8a71a7a5ce3cc5831781afaa2a459f96R1-R8):
Added configuration details for Velero, including model name, author
information, CRD URLs, and versioning.
*
[`pyproject.toml`](diffhunk://#diff-50c86b7ed8ac2cf95bd48334961bf0530cdc77b5a56f852c5c61b89d735fd711R42):
Added Velero to the list of dependencies.

Fixes #77

Signed-off-by: Sambhav Kothari <sambhavs.email@gmail.com>